opt for one top-level library
