+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# Changelog
+## 0.2.0
+
+- *BREAKING* Exposing only one top-level library: `firebase.dart`.
 
 ## 0.1.0
