+++ example/auth/index.dart
@@ -1,14 +1,12 @@
 library firebase3.example.auth;
 
 import 'dart:html';
-import 'package:firebase3/firebase.dart' as firebase;
-import 'package:firebase3/auth.dart';
-import 'package:firebase3/firebase.dart';
+import 'package:firebase3/firebase.dart' as fb;
 
 // Update firebase.initializeApp() with information from your project.
 // See <https://firebase.google.com/docs/web/setup>.
 void main() {
-  firebase.initializeApp(
+  fb.initializeApp(
       apiKey: "TODO",
       authDomain: "TODO",
       databaseURL: "TODO",
@@ -18,7 +16,7 @@ void main() {
 }
 
 class AuthApp {
-  Auth auth;
+  fb.Auth auth;
   FormElement registerForm;
   InputElement email, password;
   AnchorElement logout;
@@ -26,7 +24,7 @@ class AuthApp {
   ParagraphElement error;
 
   AuthApp() {
-    this.auth = firebase.auth();
+    this.auth = fb.auth();
     this.email = querySelector("#email");
     this.password = querySelector("#password");
     this.authInfo = querySelector("#auth_info");
@@ -61,12 +59,12 @@ class AuthApp {
 
     // When auth state changes
     auth.onAuthStateChanged.listen((e) {
-      User u = e.user;
+      fb.User u = e.user;
       _setLayout(u);
     });
   }
 
-  void _setLayout(User user) {
+  void _setLayout(fb.User user) {
     if (user != null) {
       this.registerForm.style.display = "none";
       this.logout.style.display = "block";
