+++ example/realtime_database/index.dart
@@ -1,13 +1,12 @@
 library firebase3.example.realtime_database;
 
 import 'dart:html';
-import 'package:firebase3/firebase.dart' as firebase;
-import 'package:firebase3/database.dart';
+import 'package:firebase3/firebase.dart' as fb;
 
 // Update firebase.initializeApp() with information from your project.
 // See <https://firebase.google.com/docs/web/setup>.
 void main() {
-  firebase.initializeApp(
+  fb.initializeApp(
       apiKey: "TODO",
       authDomain: "TODO",
       databaseURL: "TODO",
@@ -17,14 +16,14 @@ void main() {
 }
 
 class MessagesApp {
-  Database database;
-  DatabaseReference ref;
+  fb.Database database;
+  fb.DatabaseReference ref;
   UListElement messages;
   InputElement newMessage;
   FormElement newMessageForm;
 
   MessagesApp() {
-    this.database = firebase.database();
+    this.database = fb.database();
     this.ref = database.ref("messages");
     this.messages = querySelector("#messages");
     this.newMessage = querySelector("#new_message");
@@ -43,7 +42,7 @@ class MessagesApp {
 
   void showMessages() {
     this.ref.onChildAdded.listen((e) {
-      DataSnapshot datasnapshot = e.snapshot;
+      fb.DataSnapshot datasnapshot = e.snapshot;
 
       var spanElement = new SpanElement()..text = datasnapshot.val()["text"];
 
@@ -70,7 +69,7 @@ class MessagesApp {
     });
 
     this.ref.onChildChanged.listen((e) {
-      DataSnapshot datasnapshot = e.snapshot;
+      fb.DataSnapshot datasnapshot = e.snapshot;
       var element = querySelector("#${datasnapshot.key} span");
 
       if (element != null) {
@@ -79,7 +78,7 @@ class MessagesApp {
     });
 
     this.ref.onChildRemoved.listen((e) {
-      DataSnapshot datasnapshot = e.snapshot;
+      fb.DataSnapshot datasnapshot = e.snapshot;
 
       var element = querySelector("#${datasnapshot.key}");
 
@@ -89,7 +88,7 @@ class MessagesApp {
     });
   }
 
-  void _deleteItem(DataSnapshot datasnapshot) {
+  void _deleteItem(fb.DataSnapshot datasnapshot) {
     this
         .ref
         .child(datasnapshot.key)
@@ -97,7 +96,7 @@ class MessagesApp {
         .catchError((e) => print("Error while deleting item, $e"));
   }
 
-  void _uppercaseItem(DataSnapshot datasnapshot) {
+  void _uppercaseItem(fb.DataSnapshot datasnapshot) {
     var value = datasnapshot.val();
     var valueUppercase = value["text"].toString().toUpperCase();
     value["text"] = valueUppercase;
