+++ lib/src/interop/app_interop.dart
@@ -1,11 +1,12 @@
 @JS('firebase.app')
 library firebase3.app_interop;
 
-import 'package:firebase3/src/interop/auth_interop.dart';
-import 'package:firebase3/src/interop/database_interop.dart';
-import 'package:firebase3/src/interop/firebase_interop.dart';
-import 'package:firebase3/src/interop/storage_interop.dart';
 import 'package:js/js.dart';
+
+import 'auth_interop.dart';
+import 'database_interop.dart';
+import 'firebase_interop.dart';
+import 'storage_interop.dart';
 
 @JS('App')
 abstract class AppJsImpl {
