+++ lib/src/utils.dart
@@ -1,11 +1,10 @@
-library firebase3.utils;
-
 import 'dart:async';
 import 'dart:convert';
 import 'dart:js';
 
-import 'package:firebase3/src/interop/js_interop.dart' as js;
 import 'package:js/js.dart';
+
+import 'interop/js_interop.dart' as js;
 
 /// Returns Dart representation from JS Object.
 dynamic dartify(jsObject) {
